requirement.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.5.0.5
celery==4.1.1
certifi==2020.4.5.1
chardet==3.0.4
Django==1.8.2
django-haystack==2.7.0
django-redis==4.7.0
django-tinymce==3.0.1
fdfs-client-py==1.2.6
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
kombu==4.6.8
mutagen==1.44.0
Pillow==7.1.1
PyMySQL==0.9.3
pytz==2019.3
redis==3.5.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
uWSGI==2.0.18
vine==1.3.0
Whoosh==2.7.4
zipp==3.1.0
